add simple tests

+++ database/migrations/2017_05_06_003659_create_companies_table.php
@@ -19,7 +19,7 @@ class CreateCompaniesTable extends Migration
             $table->string('identifier');
             $table->integer('shares');
             $table->integer('free_shares');
-            $table->double('value');
+            $table->integer('value');
             $table->timestamps();
         });
     }
